requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
asgiref==3.10.0
binaryornot==0.4.4
bitarray==3.7.2
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cookiecutter==2.6.0
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
Django==5.2.7
django-allauth==65.13.0
django-appconf==1.1.0
django-braces==1.17.0
django-browser-reload==1.21.0
django-compressor==4.5.1
django-cors-headers==4.9.0
django-cotton==2.1.3
django-crispy-forms==2.4
django-extensions==4.1
django-jazzmin==3.0.1
django-nested-admin==4.1.4
django-phonenumber-field==8.3.0
django-select2==8.4.1
django-shortcuts==1.6
django-tailwind==4.3.1
django-viewcomponent==1.0.11
django-widget-tweaks==1.5.0
django_components==0.143.0
djangorestframework==3.16.1
djc_core_html_parser==1.0.3
docx==0.2.4
dotenv==0.9.9
fuzzywuzzy==0.18.0
greenlet==3.2.4
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jinja2-time==0.2.0
jiter==0.11.1
json5==0.12.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==2.0.1
packaging==25.0
pathspec==0.12.1
pdfreader==0.1.15
phonenumbers==9.0.17
pillow==12.0.0
platformdirs==4.5.0
poyo==0.5.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pynvim==0.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
python-monkey-business==1.1.0
python-pptx==1.0.2
python-slugify==8.0.4
python3-openid==3.2.0
pytokens==0.2.0
PyYAML==6.0.3
rcssmin==1.1.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rjsmin==1.2.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.0.0
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.6.0
xlsxwriter==3.2.9
